+++ lib/Util/ObjectTypes.php
@@ -91,7 +91,6 @@ class ObjectTypes
         \Stripe\SetupIntent::OBJECT_NAME => \Stripe\SetupIntent::class,
         \Stripe\ShippingRate::OBJECT_NAME => \Stripe\ShippingRate::class,
         \Stripe\Sigma\ScheduledQueryRun::OBJECT_NAME => \Stripe\Sigma\ScheduledQueryRun::class,
-        \Stripe\SKU::OBJECT_NAME => \Stripe\SKU::class,
         \Stripe\Source::OBJECT_NAME => \Stripe\Source::class,
         \Stripe\SourceTransaction::OBJECT_NAME => \Stripe\SourceTransaction::class,
         \Stripe\Subscription::OBJECT_NAME => \Stripe\Subscription::class,